Albot - Added more classes
