--- src/main/java/duke/Event.java
@@ -1,15 +1,0 @@
-package duke;
-
-public class Event extends Task {
-    protected String at;
-
-    public Event(String description, String at) {
-        super(description);
-        this.at = at;
-    }
-
-    @Override
-    public String toString() {
-        return "[E]" + super.toString() + " (at: " + at + ")";
-    }
-}
