--- src/main/java/duke/Todo.java
@@ -1,12 +1,0 @@
-package duke;
-
-public class Todo extends Task {
-    public Todo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
