+++ src/main/java/duke/task/Event.java
@@ -1,0 +1,15 @@
+package duke.task;
+
+public class Event extends Task {
+    public String at;
+
+    public Event(String description, String at) {
+        super(description);
+        this.at = at;
+    }
+
+    @Override
+    public String toString() {
+        return "[E]" + super.toString() + " (at: " + at + ")";
+    }
+}
